Refactor: Update storage import paths and type imports
