+++ server/services/xrpc/services/push-notification-service.ts
@@ -4,7 +4,7 @@
  */
 
 import type { Request, Response } from 'express';
-import { storage } from '../../storage';
+import { storage } from '../../../storage';
 import { requireAuthDid } from '../utils/auth-helpers';
 import { handleError } from '../utils/error-handler';
 import { registerPushSchema, unregisterPushSchema } from '../schemas';
