+++ server/services/xrpc/services/starter-pack-service.ts
@@ -4,7 +4,7 @@
  */
 
 import type { Request, Response } from 'express';
-import { storage } from '../../storage';
+import { storage } from '../../../storage';
 import { handleError } from '../utils/error-handler';
 import { resolveActor } from '../utils/resolvers';
 import { transformBlobToCdnUrl } from '../utils/serializers';
